--- a/addons/statemachine/XEH_PREP.hpp
+++ b/addons/statemachine/XEH_PREP.hpp
@@ -9,3 +9,5 @@ PREP(getCurrentState);
 PREP(manualTransition);
 PREP(toString);
 PREP(updateList);
+
+PREP(dev_dumpPerformanceCounters);