--- a/addons/statemachine/script_component.hpp
+++ b/addons/statemachine/script_component.hpp
@@ -2,6 +2,7 @@
 #include "\x\cba\addons\main\script_mod.hpp"
 
 // #define DISABLE_COMPILE_CACHE
+// #define STATEMACHINE_PERFORMANCE_COUNTERS
 
 #ifdef DEBUG_ENABLED_STATEMACHINE
     #define DEBUG_MODE_FULL