@BaerMitUmlaut 2017-05-07
Two things:
- Why not save the counter on the state machine itself instead of in a GVAR?
- Personally I'd not use another prefix for the function, the name of the function is almost 50 characters long, I'd rather mention in the header that this is only for development purposes