@PabstMirror 2017-05-07
1. Can still track removed state machine and what their performance was
2. Fixed